Base: write an 8-bit integer as a number, like every other width
OutputStream's text mode wrote int8_t and uint8_t through std::ostream,
which spells them as the character they stand for, while InputStream has
always read them back as numbers. The two were not inverses, so nothing
could round trip an 8-bit value outside binary mode: a field mask of 3
went out as an unprintable control character and came back as whatever
the digits after it happened to parse to, taking the rest of the stream
with it.

Found by the material list's per field doc file, which writes a mask
byte and then values that have to line up behind it. Every entry past
the mask came back wrong, in the text mode a document saved to a
directory uses.

Widened the writer rather than narrowing the reader: the number is what
all six other integer widths in this class write, and it is what the
reader already expected. Nothing in the tree wrote an 8-bit integer
through this class -- it could not have worked -- so there are no files
in the old shape to keep reading.

tests/src/Base/Stream.cpp round trips every width, both bools, both
floats and a run of bytes behind a count, in binary and text. The
material list suite is the end-to-end half: it goes back to a uint8_t
mask here, which is the exact code that failed before this.

+++ src/Base/Stream.h
@@ -100,13 +100,20 @@ public:
         return *this;
     }
 
+    // ⚠️ The text form of an 8-bit integer is its number, not the character
+    // it stands for. std::ostream writes int8_t and uint8_t as characters
+    // because they are signed char and unsigned char, while InputStream has
+    // always read them back as numbers -- so the two were not inverses and
+    // nothing could round trip through them outside binary mode. Widened
+    // here rather than narrowed there, because the number is what every
+    // other integer width in this class writes.
     OutputStream& operator<<(int8_t ch)
     {
         if(_binary) {
             _out.write((const char*)&ch, sizeof(int8_t));
         }
         else {
-            _out << ch << '\n';
+            _out << static_cast<int>(ch) << '\n';
         }
         return *this;
     }
@@ -117,7 +124,7 @@ public:
             _out.write((const char*)&uch, sizeof(uint8_t));
         }
         else {
-            _out << uch << '\n';
+            _out << static_cast<unsigned>(uch) << '\n';
         }
         return *this;
     }
